revisit: Drop backlogged frames if there are too many queued
This patch is a recorder-controllable version of the queue-cutting
idea that was discussed for a long time. While there are better
fixes available for the problem, I keep this code in my instrumented
tree in order to be able to see the effect of developer-controlled
forced queue cuts, e.g. on the recovery of the picture (depending
on key frames) or on CPU load.

This commit can probably be dropped if this branch is to be
put in production.