recorder: Display individual outgoing metrics as they arrive
This is a bit of additonal code, probably to be removed once I
understand what is going on well enough. But for now, it makes
it possible to graph individual incoming metrics and make sure
we get them at the right time and with the right values.

Unlike the server-side code, this one lets us see the accumulated
data as well