Activate metrics when receiving metrics message
Metrics are similar to the existing reports, but are finer grained.
Each metric is sent using a tagged message, and metrics can be sent
independently from one another (unlike reports, which group several
statistics in a single message). This makes it easier to send metrics
at the time the data becomes available.

Using a tagged metric system also means it's possible to add new
metrics without necesssarily bumping the protocol. As a rule, metrics
are a "quality of service" aspect only, so unknown metric IDs can
safely be ignored, and not sending a metric should only impact the
fine-tuning of the encoding, not break the communication, meaning
that client and servers are compatible irrespective of the metrics
they send or receive.

Note: The fact that a protocol update is not *required* does not mean
that one can be careless about metric IDs. In particular, client and
server should agree on the metric IDs they both recognize. Sending a
metric with a mismatched ID would lead the server to interpret the data
incorrectly. So metric IDs are themselves part of the protocol, and
new metrics should be added at the end of the list.