Add metric for number of bytes displayed per second
This metrics shows the number of (encoded) bytes displayed per
second. If there is a discrepancy between that and the number of bytes
decoded, we know that the bottleneck is the display stage. Note that
in the case of overlay (full hardware decoding), we assume that
putting things in the decoder mean they are immediately displayed.