Add 'decoded' metrics, number of bytes and frames decoded
Once the decoder has done its job and schedules frames for display,
we track how many frames were decoded and how many bytes this represents.
We use the original 'encoded' size in order to be able to do
size comparisons in the server and be able to detect which stage
is the bottleneck.