gstreamer: Improve the statistics collection
By the time schedule_frame() pulls a sample off GStreamer's pipeline the
frame may have been decoded for hundreds of milliseconds, making the
decoding time all but meaningless.
This patch ensures the statistics are always collected by
sink_event_probe() which is called as soon as the decoded frame is
available.
Note that even so the decoding time may be overestimated as the frame
may have been waiting for a while in encoded form for a spot to be freed
in the GStreamer pipeline's sink queue.

Signed-off-by: Francois Gouget <fgouget@codeweavers.com>
Acked-by: Frediano Ziglio <fziglio@redhat.com>