channel-display: Rename the frame mmtime variables
The point of the mmtime timestamps is that they are the same on the
server and client thanks to the client running its own mmtime clock
synchronized, modulo a server-controlled offset, to the server's
mmtime clock.
So the frame mmtime timestamps are neither tied to the server nor the
client. They are however tied to the frame.

Signed-off-by: Francois Gouget <fgouget@codeweavers.com>