usb-redir: move USB events handling to USB backend
Before this commit:
usb-device-manager starts thread for handling libusb events:
on Linux - from the beginning (as it is needed for hotplug
callbacks), on Windows - starts it on first redirection and
stops when there are no redirections (it can't keep the thread
when there are no redirections as with libusb < 1.0.21 it will
not be able to force the thread to exit if there are no events).

Current commit moves the event thread and handling events
completely to usb backend; usb-device-manager and other
are not aware of libusb and should not assume what it needs
to work. We start the event thread from the beginning on both
Linux and Windows. On Linux it works only for hotplug callbacks,
on Windows - just waits until device redirection starts.
On dispose of usb-device-manager (when hotplug callbacks are
deregistered), we interrupt the thread once to stop it.

This removes many lines of code and also removes all the
differences between Linux and Windows in usb-device-manager.

Signed-off-by: Yuri Benditovich <yuri.benditovich@daynix.com>
Acked-by: Victor Toso <victortoso@redhat.com>